requirements.txt
asgiref==3.3.1
backcall==0.2.0
certifi==2020.12.5
cffi==1.14.4
decorator==4.4.2
Django==3.1.6
django-elasticsearch-dsl==7.1.4
django-elasticsearch-dsl-drf==0.21
django-nine==0.2.4
djangorestframework==3.12.2
elasticsearch==7.10.1
elasticsearch-dsl==7.3.0
ipaddress==1.0.23
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.0
numpy==1.19.5
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.14
ptyprocess==0.7.0
PyAudio==0.2.11
pycparser==2.20
Pygments==2.7.4
python-dateutil==2.8.1
pyttsx3==2.90
pytz==2021.1
scipy==1.5.4
six==1.15.0
sounddevice==0.4.1
SpeechRecognition==3.8.1
sqlparse==0.4.1
traitlets==4.3.3
urllib3==1.26.3
uWSGI==2.0.19.1
wcwidth==0.2.5
